requirements.txt
# Core dependencies
feedparser==6.0.11
python-dateutil==2.9.0
pytz==2024.1

# HTML processing
beautifulsoup4==4.12.3
lxml==5.1.0
readability-lxml==0.8.4.1

# HTTP requests
requests==2.32.3
aiohttp==3.9.3

# Web framework
Flask[async]==3.0.0
Flask-Caching==2.1.0

# Async
asyncio==3.4.3